break helper functions into separate module
